fix(iva): castear $5::numeric en upsert de costos
El parametro $5 se usaba en COALESCE($5,21) (integer) y en
COALESCE($5, product_costs.alicuota_iva, 21) (numeric); Postgres deducia tipos
inconsistentes para $5 y tiraba "inconsistent types deduced for parameter $5" en
CADA guardado de costos (import de Excel incluido). Se castea $5::numeric en ambos
lugares para fijar el tipo.

+++ server.js
@@ -4447,10 +4447,10 @@ app.post('/api/reporte/costos', requireAuth, async (req, res) => {
       const alic = ALIC_VALIDAS.includes(alicRaw) ? alicRaw : null;
       await pool.query(`
         INSERT INTO product_costs (client_id, mla_id, title, costo_unit, alicuota_iva, notas, updated_at)
-        VALUES ($1,$2,$3,$4,COALESCE($5,21),$6,NOW())
+        VALUES ($1,$2,$3,$4,COALESCE($5::numeric,21),$6,NOW())
         ON CONFLICT (client_id, mla_id) DO UPDATE SET
           title=$3, costo_unit=$4,
-          alicuota_iva=COALESCE($5, product_costs.alicuota_iva, 21),
+          alicuota_iva=COALESCE($5::numeric, product_costs.alicuota_iva, 21),
           notas=$6, updated_at=NOW()
       `, [client_id, c.mla_id, c.title, c.costo_unit||0, alic, c.notas||'']);
     }
